[ios] Link PlatformView back to semantics tree (#46471)
The PlatformView does not have a semantics container when added to semantics tree, this PR gives it a semantics container to ensure accessibility traversal works.

This fixes https://github.com/flutter/flutter/issues/135504, which is a regression of https://github.com/flutter/engine/commit/7626a848410cc3523e5cb9ea9ebbf7765c8c664b

Before https://github.com/flutter/engine/commit/7626a848410cc3523e5cb9ea9ebbf7765c8c664b, the traversal works because the PlatformView is added to the accessibilityElements of the FlutterPlatFormViewSemanticsContainer, which implicitly made the FlutterPlatFormViewSemanticsContainer as the PlatformVIew's AccessibilityContainer. 

Now we use the PlatformVIew as the nativeAccessibility of the  FlutterPlatFormViewSemanticsContainer, we need to expilicitly set the container. 

This needs to be cherry-picked since the commit caused the regression was cherry-picked in https://github.com/flutter/engine/commit/7626a848410cc3523e5cb9ea9ebbf7765c8c664b

[C++, Objective-C, Java style guides]: https://github.com/flutter/engine/blob/main/CONTRIBUTING.md#style